kubeadm: refactor KubernetesReleaseVersion()
KubernetesReleaseVersion() is a recursive function with side effects which
does more things than required by some callers and as such is difficult to
reason about.

Rather move its code to the only suitable consumer which is
`SetClusterDynamicDefaults()` and reorganize it to avoid endless
recursions like in the case when a faulty remote server responses
with "stable-1" for the request "stable" and with "stable" for
the request "stable-1". The new logic is as follows:

SetClusterDynamicDefaults() accept a ClusterConfiguration whose
KubernetesVersion field may be

1. a version or
2. just a version label supposed to be resolved remotely into a version or
3  an empty string.

In the first case the only task would be to canonicize the version
with prefixed "v". In the second case the label would be resolved
remotely. And in the third case we could try to read the gitVersion
and resort to kubeadmapiv1beta1.DefaultKubernetesVersion (which is "stable-1"
currently) if it's not set.

After SetClusterDynamicDefaults() has been called it's guarantied
that ClusterConfiguration.KubernetesVersion is set and valid.